// src/app/modules/openapi/model/update-profile-dto.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateProfileDTO { 
    userName: string;
    email: string;
    job?: string;
    creditNo?: string;
    creditLimit?: number;
    phone?: string;
}